esp32/python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.59.1
    # via matplotlib
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.5
    # via -r requirements.in
numpy==2.3.2
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==25.0
    # via matplotlib
pillow==11.3.0
    # via matplotlib
pyparsing==3.2.3
    # via matplotlib
pyserial==3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via matplotlib
six==1.17.0
    # via python-dateutil
